+++ app/cdap/components/ColumnView/components/ColumnViewWidget/DrawerWidgetHeading/__test__/DrawerWidgetHeading.test.tsx
@@ -14,15 +14,14 @@
  * the License.
  */
 
-import { prepareDataQualtiy } from 'components/ColumnView/SelectColumnsList';
-import {
-  mockColumnData,
-  mockDataQuality,
-  mockResult,
-} from 'components/ColumnView/__test__/ColumnView.test';
+import { render, screen } from '@testing-library/react';
+import React from 'react';
+import DrawerWidgetHeading from 'components/ColumnView/components/ColumnViewWidget/DrawerWidgetHeading';
 
-describe('It Should test the utils function prepareDataQualtiy', () => {
-  it('should test whether the utils function return the expected Value', () => {
-    expect(prepareDataQualtiy(mockDataQuality, mockColumnData)).toStrictEqual(mockResult);
+describe('It Should test DrawerWidgetHeading Component', () => {
+  it('Should render DrawerWidgetheading Component and test the column view heading text', () => {
+    render(<DrawerWidgetHeading headingText={'Column View Panel'} />);
+    const headingText = screen.getByTestId('drawer-widget-heading-text');
+    expect(headingText).toHaveTextContent('Column View Panel');
   });
 });
