+++ app/cdap/components/ColumnView/components/DataQualityCircularProgressBar/index.tsx
@@ -18,7 +18,7 @@ import { Box } from '@material-ui/core';
 import React from 'react';
 import SemiCircleProgressBar from 'react-progressbar-semicircle';
 import styled from 'styled-components';
-import { IDataQualityRecord } from 'components/ColumnView/SelectColumnsList/index';
+import { IDataQualityRecord } from 'components/ColumnView/components/SelectColumnsList/index';
 
 export interface IDataQualityCircularProgressBarProps {
   dataQualityPercentValue: number;
