+++ app/cdap/components/ColumnView/components/SelectColumnsList/__test__/SelectColumnsList.test.tsx
@@ -15,7 +15,7 @@
  */
 
 import { render, screen } from '@testing-library/react';
-import SelectColumnsList from 'components/ColumnView/SelectColumnsList';
+import SelectColumnsList from 'components/ColumnView/components/SelectColumnsList';
 import { mockColumnData, mockDataQuality } from 'components/ColumnView/__test__/ColumnView.test';
 import T from 'i18n-react';
 import React from 'react';
